prima: ignore false positive array bounds warnings
A false postive array bounds warning is thrown near line 1203 of this file
on GCC 4.8 and it's decendants. The associated bug in GCC has been fixed in
newer toolchains (such as the base for Linaro 4.9.3), but the default
toolchain for CyanogenMod (4.8) still has this bug. So let's disable the
warning for this file to avoid headaches and forcing people to use a newer
compiler.

Change-Id: I3a7db938cf06ba8b8ee299331fb854e3ba0674fa